// tools/ipv6-ula-generator/types.ts
/* v8 ignore file -- type-only module */

type Ipv6UlaMessagesCatalog = Readonly<{
  generatorTitle: string
  generatorDescription: string
  sitePrefixLabel: string
  generatingLabel: string
  regenerateLabel: string
  copySitePrefixLabel: string
  copyGlobalIdLabel: string
  copySelectedSubnetLabel: string
  copyFirstSubnetLabel: string
  copyLastSubnetLabel: string
  copiedLabel: string
  detailsTitle: string
  detailsDescription: string
  globalIdLabel: string
  ulaSpaceLabel: string
  ulaSpaceValue: string
  localBitLabel: string
  localBitValue: string
  prefixLengthLabel: string
  prefixLengthValue: string
  randomSourceLabel: string
  randomSourceValue: string
  subnetCapacityLabel: string
  layoutTitle: string
  layoutDescription: string
  localPrefixSegmentLabel: string
  globalIdSegmentLabel: string
  subnetIdSegmentLabel: string
  interfaceIdSegmentLabel: string
  bitsLabel: string
  plannerTitle: string
  plannerDescription: string
  subnetIdLabel: string
  subnetIdDescription: string
  subnetIdPlaceholder: string
  subnetIdInvalid: string
  selectedSubnetLabel: string
  boundariesTitle: string
  boundariesDescription: string
  firstSubnetLabel: string
  lastSubnetLabel: string
  privacyTitle: string
  privacyDescription: string
  reuseTitle: string
  reuseDescription: string
  errorTitle: string
  cryptoUnavailableMessage: string
}>

type Ipv6UlaMessages = Ipv6UlaMessagesCatalog & {
  meta: {
    name: string
    description: string
  }
}

export type { Ipv6UlaMessages, Ipv6UlaMessagesCatalog }